refactor(discord): bootstrap のランタイムヘルパを runtime.ts へ分離し旧テストを再配置する
